@@ -4,11 +4,9 @@
       <WV2WinRTExplicitIncludesOnly>True</WV2WinRTExplicitIncludesOnly>
     </PropertyGroup>
     <PropertyGroup>
-        <TargetFramework>net8.0-windows</TargetFramework>
+        <TargetFramework>net9.0-windows</TargetFramework>
         <OutputType>WinExe</OutputType>
 		<UseWpf>true</UseWpf>
-
-        <!-- Project Options -->
         <Nullable>enable</Nullable>
         <LangVersion>latest</LangVersion>
         <ImplicitUsings>enable</ImplicitUsings>
@@ -19,6 +17,12 @@
         <PackageReleaseNotes>RELEASE.md</PackageReleaseNotes>
         <PackageReadmeFile>README.md</PackageReadmeFile>
         <AnalysisLevel>latest</AnalysisLevel>
+        <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
+        <!-- MSIX Packaging -->
+        <WindowsPackageType>None</WindowsPackageType>
+        <EnableMsixTooling>true</EnableMsixTooling>
+        <AppxBundle>Never</AppxBundle>
+        <GenerateAppInstallerFile>false</GenerateAppInstallerFile>
     </PropertyGroup>
 
 
@@ -33,13 +37,20 @@
     </ItemGroup>
 
     <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="9.0.8" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.8" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="9.0.8" />
         <PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.*" />
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
-
-     
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.8" />
     </ItemGroup>
 
     <ItemGroup>
+      <EmbeddedResource Include="Resources\Strings.resx" />
+      <EmbeddedResource Include="Resources\Strings.es.resx" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <None Include="Package\ClientApp.Package.wapproj" />
       <None Include="README.md">
         <Pack>True</Pack>
         <PackagePath>\</PackagePath>
@@ -47,5 +58,10 @@
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\CorePolicyEngine\CorePolicyEngine.csproj" />
+        <ProjectReference Include="..\Security\Security.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+      <Folder Include="Logging\" />
+      <Folder Include="Resources\" />
     </ItemGroup>
 </Project>
