@@ -5,33 +5,38 @@
 // License: MIT
 // Do not remove file headers
 
-
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using System.IO; // for Path & Directory
 using System.Windows.Input;
+using CorePolicyEngine.Parsing;
+using CorePolicyEngine.AdminTemplates;
+using CorePolicyEngine; // for Result
+using Microsoft.Extensions.Logging;
+using ClientApp.Logging; // for source-generated logging extension methods
+using CorePolicyEngine.Storage;
 
 namespace ClientApp.ViewModels;
 
 public class PolicyEditorViewModel : INotifyPropertyChanged
 {
-    private readonly IAdmxCatalogLoader _loader;
+    private readonly IAdminTemplateLoader _loader;
+    private readonly ILogger<PolicyEditorViewModel> _logger;
+    private readonly IAuditWriter _audit;
 
-    public ObservableCollection<AdmxCategory> Categories { get; } = [];
-    public ObservableCollection<AdmxPolicy> Policies { get; } = [];
-    public ObservableCollection<AdmxPolicy> FilteredPolicies { get; } = []; // retained
-    public ObservableCollection<PolicySettingViewModel> CurrentSettings { get; } = [];
     public ObservableCollection<CategoryTreeItem> CategoryTree { get; } = [];
+    public ObservableCollection<PolicySummary> Policies { get; } = [];
+    public ObservableCollection<PolicySummary> FilteredPolicies { get; } = [];
+    public ObservableCollection<PolicySettingViewModel> SelectedPolicySettings { get; } = [];
 
-    private AdmxCatalog? _catalog;
-    public AdmxCatalog? Catalog { get => _catalog; private set { _catalog = value; OnPropertyChanged(); } }
+    private AdminTemplateCatalog? _catalog;
+    public AdminTemplateCatalog? Catalog { get => _catalog; private set { _catalog = value; OnPropertyChanged(); } }
 
-    private AdmxPolicy? _selectedPolicy;
-    public AdmxPolicy? SelectedPolicy { get => _selectedPolicy; set { _selectedPolicy = value; OnPropertyChanged(); LoadSettingsForSelected(); } }
+    private PolicySummary? _selectedPolicy;
+    public PolicySummary? SelectedPolicy { get => _selectedPolicy; set { if (_selectedPolicy != value) { _selectedPolicy = value; OnSelectedPolicyChanged(); OnPropertyChanged(); } } }
 
-    private AdmxCategory? _selectedCategory;
-    public AdmxCategory? SelectedCategory { get => _selectedCategory; set { if (_selectedCategory != value) { _selectedCategory = value; OnPropertyChanged(); SelectedPolicy = null; } } }
+    private CategoryTreeItem? _selectedCategoryNode;
+    public CategoryTreeItem? SelectedCategoryNode { get => _selectedCategoryNode; set { if (_selectedCategoryNode != value) { _selectedCategoryNode = value; OnPropertyChanged(); SelectedPolicy = null; } } }
 
     private string? _lastSearch;
     private string? _searchText;
@@ -44,7 +49,8 @@ public class PolicyEditorViewModel : INotifyPropertyChanged
             {
                 _searchText = value; OnPropertyChanged();
                 ApplySearchFilter(_searchText);
-                BuildCategoryTree();
+                RebuildCategoryRoots();
+                _logger.SearchFilterApplied(_searchText ?? string.Empty);
             }
         }
     }
@@ -52,133 +58,148 @@ public class PolicyEditorViewModel : INotifyPropertyChanged
     public ICommand LoadPoliciesCommand { get; }
     public ICommand SearchLocalPoliciesCommand { get; }
 
-    public PolicyEditorViewModel(IAdmxCatalogLoader loader)
+    public PolicyEditorViewModel(IAdminTemplateLoader loader, ILogger<PolicyEditorViewModel> logger, IAuditWriter audit)
     {
         _loader = loader;
-        LoadPoliciesCommand = new RelayCommand(async _ => await LoadDefaultSubsetAsync(), _ => true);
-        SearchLocalPoliciesCommand = new RelayCommand(async _ => await SearchLocalPoliciesAsync(SearchText, CancellationToken.None), _ => true);
+        _logger = logger;
+        _audit = audit;
+        LoadPoliciesCommand = new RelayCommand(async _ => await LoadDefaultSubsetAsync().ConfigureAwait(false), _ => true);
+        SearchLocalPoliciesCommand = new RelayCommand(async _ => await LoadEntireCatalogAsync("en-US", CancellationToken.None).ConfigureAwait(false), _ => true);
+        _logger.Initialized(); // TODO localize message key via resources
     }
 
     private async Task LoadDefaultSubsetAsync()
     {
-        string winDir = Environment.GetFolderPath(Environment.SpecialFolder.Windows);
-        string policyDef = Path.Combine(winDir, "PolicyDefinitions");
-        if (Directory.Exists(policyDef))
-        {
-            List<string> admxFiles = Directory.GetFiles(policyDef, "*.admx").Take(25).ToList();
-            await LoadCatalogAsync(admxFiles, "en-US", CancellationToken.None);
-        }
+        await LoadEntireCatalogAsync("en-US", CancellationToken.None).ConfigureAwait(false);
     }
 
-    public async Task LoadCatalogAsync(IEnumerable<string> paths, string? culture, CancellationToken token)
+    private async Task LoadEntireCatalogAsync(string languageTag, CancellationToken token)
     {
-        Result<AdmxCatalog> result = await _loader.LoadAsync(paths.ToList(), culture, token);
-        if (!result.Success)
+        var sw = System.Diagnostics.Stopwatch.StartNew();
+        Result<AdminTemplateCatalog> result;
+        try
         {
+            result = await this._loader.LoadLocalCatalogAsync(languageTag, 50, token).ConfigureAwait(false); // cap initial load
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "Unexpected exception loading catalog language {Language}", languageTag); // TODO localize
             return;
         }
-
+        sw.Stop();
+        if (!result.Success || result.Value is null)
+        {
+            _logger.CatalogLoadFailed(languageTag);
+            return;
+        }
         Catalog = result.Value;
-        Categories.Clear(); foreach (AdmxCategory? c in Catalog?.Categories.OrderBy(c => c.Name)!)
-        {
-            Categories.Add(c);
-        }
-
-        Policies.Clear(); foreach (AdmxPolicy? p in Catalog.Policies.OrderBy(p => p.Name))
-        {
-            Policies.Add(p);
-        }
-
+        Policies.Clear();
+        foreach (var s in Catalog.Summaries.OrderBy(s => s.DisplayName)) Policies.Add(s);
         ApplySearchFilter(_lastSearch);
-        BuildCategoryTree();
+        RebuildCategoryRoots();
+        _logger.CatalogLoaded(languageTag, Policies.Count, sw.ElapsedMilliseconds);
     }
 
     public async Task SearchLocalPoliciesAsync(string? query, CancellationToken token)
     {
-        string winDir = Environment.GetFolderPath(Environment.SpecialFolder.Windows);
-        string policyDef = Path.Combine(winDir, "PolicyDefinitions");
-        if (Directory.Exists(policyDef))
-        {
-            List<string> allFiles = Directory.GetFiles(policyDef, "*.admx").ToList();
-            if (Catalog == null || Catalog.Policies.Count < allFiles.Count)
-            {
-                await LoadCatalogAsync(allFiles, "en-US", token);
-            }
-        }
+        if (Catalog == null) await LoadEntireCatalogAsync("en-US", token).ConfigureAwait(false);
         ApplySearchFilter(query);
-        BuildCategoryTree();
+        RebuildCategoryRoots();
+        _logger.SearchExecuted(query ?? string.Empty, FilteredPolicies.Count);
+        if (!string.IsNullOrWhiteSpace(query))
+        {
+            try { await _audit.PolicyViewedAsync("SEARCH:" + query, token); } catch { }
+        }
     }
 
     public void ApplySearchFilter(string? query)
     {
         _lastSearch = query;
         FilteredPolicies.Clear();
-        IEnumerable<AdmxPolicy> source = Policies;
+        IEnumerable<PolicySummary> source = Policies;
         if (!string.IsNullOrWhiteSpace(query))
         {
             string q = query.Trim().ToLowerInvariant();
-            source = source.Where(p => (p.Name?.ToLowerInvariant().Contains(q) ?? false) || (p.Id?.ToLowerInvariant().Contains(q) ?? false));
+            source = source.Where(p =>
+                (p.DisplayName?.ToLowerInvariant().Contains(q) ?? false) ||
+                (p.Key.Name.ToLowerInvariant().Contains(q)) ||
+                (p.CategoryPath?.ToLowerInvariant().Contains(q) ?? false));
         }
-        foreach (AdmxPolicy p in source)
-        {
-            FilteredPolicies.Add(p);
-        }
+        foreach (var p in source) FilteredPolicies.Add(p);
     }
 
-    private void BuildCategoryTree()
+    private void RebuildCategoryRoots()
     {
         CategoryTree.Clear();
-        if (Catalog == null)
+        if (Catalog == null) return;
+
+        var localizedNames = new Dictionary<CategoryId, string>();
+        foreach (var admx in Catalog.AdmxDocuments)
         {
-            return;
+            foreach (var cat in admx.Categories)
+                localizedNames[cat.Id] = cat.Id.Value; // TODO: replace with localized lookup
         }
 
-        Dictionary<string, CategoryTreeItem> catLookup = Catalog.Categories.ToDictionary(c => c.Id, c => new CategoryTreeItem(c));
-        foreach (CategoryTreeItem? cat in catLookup.Values)
-        {
-            if (!string.IsNullOrWhiteSpace(cat.Category!.ParentId) && catLookup.TryGetValue(cat.Category.ParentId!, out CategoryTreeItem? parent))
-            {
-                parent.Children.Add(cat);
-            }
-        }
-        ObservableCollection<AdmxPolicy> policySource = FilteredPolicies.Any() || !string.IsNullOrWhiteSpace(_lastSearch) ? FilteredPolicies : Policies;
-        foreach (AdmxPolicy p in policySource)
-        {
-            if (catLookup.TryGetValue(p.CategoryId, out CategoryTreeItem? catNode))
-            {
-                catNode.Children.Add(new CategoryTreeItem(p));
-            }
-        }
-        foreach (CategoryTreeItem? cat in catLookup.Values.Where(c => string.IsNullOrWhiteSpace(c.Category!.ParentId)))
-        {
-            CategoryTree.Add(cat);
-        }
+        var roots = Catalog.AdmxDocuments.SelectMany(d => d.Categories)
+            .Where(c => c.Parent is null)
+            .GroupBy(c => c.Id.Value)
+            .Select(g => g.First())
+            .OrderBy(c => c.Id.Value, StringComparer.OrdinalIgnoreCase);
+
+        foreach (var r in roots)
+            CategoryTree.Add(new CategoryTreeItem(r, localizedNames[r.Id]));
     }
 
-    private void LoadSettingsForSelected()
+    public void EnsureCategoryChildren(CategoryTreeItem node)
     {
-        CurrentSettings.Clear();
-        if (SelectedPolicy == null || Catalog == null)
+        if (Catalog == null) return;
+        if (node.IsPolicy || node.Category is null) return;
+        if (node.ChildrenMaterialized) return;
+
+        node.Children.Clear();
+
+        var childCats = Catalog.AdmxDocuments.SelectMany(d => d.Categories)
+            .Where(c => c.Parent.HasValue && c.Parent.Value.Id.Value == node.Category.Id.Value)
+            .OrderBy(c => c.Id.Value, StringComparer.OrdinalIgnoreCase)
+            .ToList();
+
+        foreach (var c in childCats)
+            node.Children.Add(new CategoryTreeItem(c, c.Id.Value));
+
+        var policyMatches = Catalog.AdmxDocuments.SelectMany(d => d.Policies)
+            .Where(pol => pol.Category.Id.Value == node.Category.Id.Value)
+            .Join(Catalog.Summaries, pol => pol.Key.Name, s => s.Key.Name, (pol, s) => s)
+            .OrderBy(s => s.DisplayName, StringComparer.OrdinalIgnoreCase)
+            .ToList();
+
+        foreach (var p in policyMatches)
+            node.Children.Add(new CategoryTreeItem(p));
+
+        node.ChildrenMaterialized = true;
+        if (node.Children.Count == 0)
+            node.Children.Add(CategoryTreeItem.EmptyMarker());
+        _logger.CategoryExpanded(node.Category.Id.Value, childCats.Count, policyMatches.Count);
+    }
+
+    private async void OnSelectedPolicyChanged()
+    {
+        SelectedPolicySettings.Clear();
+        if (Catalog == null || SelectedPolicy == null) return;
+
+        var policy = Catalog.AdmxDocuments.SelectMany(d => d.Policies)
+            .FirstOrDefault(p => p.Key.Name == SelectedPolicy.Key.Name);
+        if (policy == null)
         {
+            _logger.LogWarning("Selected policy key {PolicyKey} not found in catalog documents", SelectedPolicy.Key.Name); // TODO localize
             return;
         }
 
-        if (SelectedPolicy.Parts.Any())
-        {
-            foreach (PolicyPartDefinition part in SelectedPolicy.Parts)
-            {
-                CurrentSettings.Add(new PolicySettingViewModel(SelectedPolicy.Id, part, Catalog));
-            }
-        }
-        else
-        {
-            CurrentSettings.Add(new PolicySettingViewModel(new PolicySetting(SelectedPolicy.Id, null, false, null, PolicyValueType.Boolean)));
-        }
+        foreach (var element in policy.Elements)
+            SelectedPolicySettings.Add(new PolicySettingViewModel(policy, element));
+        _logger.PolicySelected(SelectedPolicy.Key.Name, SelectedPolicySettings.Count);
+        try { await _audit.PolicySelectedAsync(SelectedPolicy.Key.Name); } catch { }
     }
 
     public event PropertyChangedEventHandler? PropertyChanged;
-    private void OnPropertyChanged([CallerMemberName] string? name = null)
-    {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-    }
+    private void OnPropertyChanged([CallerMemberName] string? name = null) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
 }
