@@ -6,14 +6,6 @@
 // Do not remove file headers
 
 
-
-// Project Name: CorePolicyEngine
-// File Name: AdmlModel.cs
-// Author: Kyle Crowder
-// Github:  OldSkoolzRoolz
-// License: MIT
-// Do not remove file headers
-
 namespace CorePolicyEngine.AdminTemplates;
 
 
@@ -41,7 +33,11 @@ public abstract record PresentationTemplate(string Id);
 // You can extend as you discover templates in the wild; keep them additive.
 public sealed record BooleanPresentation(string Id, ResourceId? Label) : PresentationTemplate(Id);
 
+
+
 public sealed record TextPresentation(string Id, ResourceId? Label, int? MaxLength) : PresentationTemplate(Id);
+
+
 
 public sealed record DecimalPresentation(string Id, ResourceId? Label, long? Min, long? Max) : PresentationTemplate(Id);
 
