@@ -6,14 +6,6 @@
 // Do not remove file headers
 
 
-
-// Project Name: CorePolicyEngine
-// File Name: ReadModel.cs
-// Author: Kyle Crowder
-// Github:  OldSkoolzRoolz
-// License: MIT
-// Do not remove file headers
-
 namespace CorePolicyEngine.AdminTemplates;
 
 
@@ -45,25 +37,42 @@ public static class Materializer
         AdmlDocument adml)
     {
         var strings = adml.StringTable;
-        Dictionary<CategoryId, string> catNames = [];
+
+        // Pre-build name + lookup maps so we can validate refs and avoid rebuilding per policy
+        var catNames = new Dictionary<CategoryId, string>();
         foreach (Category c in admx.Categories)
             catNames[c.Id] = strings.TryGetValue(c.DisplayName.Id, out var n) ? n : c.Id.Value;
 
+        var catById = new Dictionary<CategoryId, Category>();
+        foreach (Category c in admx.Categories)
+            catById[c.Id] = c;
+
         string ResolveCatPath(CategoryRef r)
         {
-            Stack<string> path = new();
-            Dictionary<CategoryId, Category> byId = [];
-            foreach (Category c in admx.Categories) byId[c.Id] = c;
+            // Treat empty / whitespace IDs as uncategorized
+            if (string.IsNullOrWhiteSpace(r.Id.Value)) return "Uncategorized";
+            if (!catById.TryGetValue(r.Id, out var cur))
+                return string.IsNullOrWhiteSpace(r.Id.Value) ? "Uncategorized" : r.Id.Value; // fallback to raw id to surface unknown category
 
-            Category cur = byId[r.Id];
+            var path = new Stack<string>();
+            int guard = 0; // cycle protection
             while (true)
             {
-                path.Push(catNames[cur.Id]);
+                if (++guard > 100)
+                {
+                    path.Push("[Cycle]");
+                    break; // protect against malformed circular parent chains
+                }
+
+                path.Push(catNames.TryGetValue(cur.Id, out var name) ? name : cur.Id.Value);
                 if (cur.Parent is null) break;
-
-                cur = byId[cur.Parent.Value.Id];
+                // If parent ref missing from index, terminate with fallback marker
+                if (!catById.TryGetValue(cur.Parent.Value.Id, out cur))
+                {
+                    path.Push("Uncategorized");
+                    break;
+                }
             }
-
             return string.Join("/", path);
         }
 
@@ -73,11 +82,22 @@ public static class Materializer
             var explain = p.ExplainText is null ? null :
                 strings.TryGetValue(p.ExplainText.Id, out var ex) ? ex : null;
 
+            string catPath;
+            try
+            {
+                catPath = ResolveCatPath(p.Category);
+            }
+            catch
+            {
+                // Absolute safety net � should no longer happen, but prevents single failure from halting enumeration
+                catPath = "Uncategorized";
+            }
+
             yield return new PolicySummary(
                 p.Key,
                 display,
                 explain,
-                ResolveCatPath(p.Category),
+                catPath,
                 p.Class,
                 p.SupportedOn?.Value,
                 ElementKinds(p.Elements),
@@ -86,7 +106,7 @@ public static class Materializer
 
         static IReadOnlyList<string> ElementKinds(IReadOnlyList<PolicyElement> elements)
         {
-            List<string> list = new(elements.Count);
+            var list = new List<string>(elements.Count);
             foreach (PolicyElement e in elements)
                 list.Add(e switch
                 {
@@ -97,7 +117,6 @@ public static class Materializer
                     EnumElement => "Enum",
                     _ => "Unknown"
                 });
-
             return list;
         }
     }
