@@ -1,11 +1,15 @@
 ﻿// Project Name: CorePolicyEngine
-// File Name: RequiredEnabledValueRule.cs (deprecated)
+// File Name: RequiredEnabledValueRule.cs
 // Author: Kyle Crowder
 // Github:  OldSkoolzRoolz
 // License: MIT
 // Do not remove file headers
 
+
 namespace CorePolicyEngine.Parsing;
 
+
 // Legacy required-value validation rule removed.
-public static class RequiredEnabledValueRuleMarker { }
+public static class RequiredEnabledValueRuleMarker
+{
+}
