@@ -1,11 +1,15 @@
 // Project Name: CorePolicyEngine
-// File Name: SqlitePolicyRepository.cs (deprecated)
+// File Name: SqlitePolicyRepository.cs
 // Author: Kyle Crowder
 // Github:  OldSkoolzRoolz
 // License: MIT
 // Do not remove file headers
 
+
 namespace CorePolicyEngine.Storage;
 
+
 // Legacy SqlitePolicyRepository removed pending redesign for new policy state persistence model.
-public static class SqlitePolicyRepositoryMarker { }
+public static class SqlitePolicyRepositoryMarker
+{
+}
