@@ -1,11 +1,15 @@
 // Project Name: CorePolicyEngine
-// File Name: RangeRule.cs (deprecated)
+// File Name: RangeRule.cs
 // Author: Kyle Crowder
 // Github:  OldSkoolzRoolz
 // License: MIT
 // Do not remove file headers
 
+
 namespace CorePolicyEngine.Validation.Rules;
 
+
 // Legacy range rule removed. Will be recreated for new DecimalElement semantics.
-public static class RangeRuleMarker { }
+public static class RangeRuleMarker
+{
+}
